+++ bin/actions/install_archive_action.js
@@ -7,12 +7,10 @@ const _ = require('lodash')
 
 const installArchiveAction = (archiveRequest, installedName) => {
   const paths = applicationConfiguration.get(`paths`)
-  // TODO cache this?
   const manifestJson = ManifestConfiguration.build(archiveRequest.stagingPath)
   const archiveFolder = _.get(archiveRequest, 'scope.installation.destination', paths.archive)
-  // TODO not sure about this resolution - maybe just the installedName
-  const archivePath = `${archiveFolder}/${installedName || manifestJson.name || archiveRequest.archive}/`
-  const installFrom = `${archiveRequest.stagingPath}${_.get(archiveRequest, 'scope.installation.subfolder', ``)}`
+  const archivePath = `${archiveFolder}/${archiveRequest.installedName}/`
+  const installFrom = `${archiveRequest.stagingPath}${_.get(archiveRequest, 'scope.installation.subfolder', '')}`
   const ignoreFolders = _.merge(
     applicationConfiguration.get(`rules.ignoreFiles`, []),
     manifestJson.ignore