+++ bin/core/terminal.js
@@ -2,7 +2,7 @@ const packageConfiguration = require('../../package')
 const _ = require('lodash')
 
 module.exports = {
-  registerVorpalCommand: (vorpal, configuration) => {
+  initializeVorpal: (vorpal, configuration) => {
     const majorVersion = parseInt(_.head(packageConfiguration.version.split(`.`)), 10)
     const applicationHandle = `${packageConfiguration.name}-${majorVersion}`